requirements.txt
fastapi==0.128.0
protobuf==6.33.5
pydantic==2.12.5
python-dotenv==1.2.1
Requests==2.32.5
uvicorn
google-generativeai